Implements Feat: ThumbnailsDraggable
